fix cpu index_elementwise

+++ paddle/phi/kernels/funcs/stride_utils.h
@@ -388,7 +388,7 @@ static inline void IndexGetStride(
   coalesce_dimensions<N>(ndim, strides_array, &stride_size, desired_shape);
 
   int num = 1;
-  for (int i = 0; i < desired_shape->size(); i++) {
+  for (size_t i = 0; i < desired_shape->size(); i++) {
     num *= (*desired_shape)[i];
   }
   *numel = num;
